backend/requirements.txt
fastapi
uvicorn
firebase-admin
groq
pydantic
python-dotenv
